+++ CHANGES.rst
@@ -7,6 +7,8 @@
 - (Hotfix) Fixed typo related to GitHub download for offline processing
 - Started modifications and support for JWST.
 - LRISr header issue
+- Limit the images written to the MasterEdges file to only the trace
+  image, mask, and detector.
 
 1.11.0 (21 Oct 2022)
 --------------------